layout: relocate the 14 post-ELK passes to src/layout/passes/ (step 2 complete)
Moves the post-ELK pass functions + their ~45 private helpers out of
layout-engine.ts into src/layout/passes/index.ts, leaving layout-engine.ts as the
ELK frontend + orchestration + LAYOUT_PIPELINE manifest (4179 -> ~1450 lines).
Two more boundary-crossers (layoutDebug/layoutEnvFlag) joined the geometry kernel.
Dependency direction is strictly one-way: passes -> geometry, never passes ->
layout-engine, so no import cycle.

alignLayerNodes + orthogonalizeEdgePoints (previously part of the public surface)
are re-exported from layout-engine.ts, so consumers are unaffected.

Pure relocation, zero behavior change. Verified byte-identical: determinism oracle
370 pass / 0 fail / 52 snapshots; full unit suite 4147 pass / 0 fail across 183
files; typecheck clean. Completes step 2 of the layout-pass-pipeline migration.
